[colorHarmonization] Fix compilation warnings

+++ src/aliceVision/colorHarmonization/GainOffsetConstraintBuilder.hpp
@@ -75,7 +75,7 @@ struct GainOffsetConstraintBuilder
     {
         // Count the number of images
         std::set<std::size_t> countSet;
-        for (int i = 0; i < _vec_relative.size(); ++i)
+        for (int i = 0; i < static_cast<int>(_vec_relative.size()); ++i)
         {
             countSet.insert(_vec_relative[i].I);
             countSet.insert(_vec_relative[i].J);
